Rename MultiSelect component and clean code with comments for cont refactoring
